Allow smithing table usage when crafting is disabled
- The listeners for both the smithing table and the anvil are now loaded regardless of the configuration. The configuration only determines if creation/upgrading is allowed, but both the anvil and the smithing table can be used for other stuff as well.
- Flattened the smithing table listener.
