+++ src/main/java/nl/pim16aap2/armoredElytra/handlers/AnvilHandler.java
@@ -24,23 +24,13 @@ import java.util.logging.Level;
 
 public class AnvilHandler extends ArmoredElytraHandler implements Listener
 {
-    protected AnvilHandler(
-        ArmoredElytra plugin,
-        boolean creationEnabled,
-        NBTEditor nbtEditor,
-        DurabilityManager durabilityManager,
-        ConfigLoader config)
-    {
-        super(plugin, creationEnabled, nbtEditor, durabilityManager, config);
-    }
-
     public AnvilHandler(
         ArmoredElytra plugin,
         NBTEditor nbtEditor,
         DurabilityManager durabilityManager,
         ConfigLoader config)
     {
-        super(plugin, true, nbtEditor, durabilityManager, config);
+        super(plugin, nbtEditor, durabilityManager, config);
     }
 
     // Handle all anvil related stuff for this plugin.
